requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
djangorestframework==3.13.1
google-auth==2.9.1
google-auth-oauthlib==0.5.2
gspread==5.4.0
httplib2==0.20.4
idna==3.3
mysqlclient==2.1.1
numpy==1.23.1
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.10
